- Atualizacao da entidade de produto para aceitar valores decimais na quantidade - Atualizacao da entidade de venda para aceitar desconto - Tratamento de desconto da venda proporcional para cada produto

+++ app/src/main/java/dev/jmoicano/salesmanager/data/entities/Product.kt
@@ -8,7 +8,7 @@ import dev.jmoicano.salesmanager.ui.data.ViewProduct
 data class Product(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val quant: Int,
+    val quant: Double,
     val unitPrice: Double,
     val saleId: Long
 )
